+++ CHANGELOG.md
@@ -22,6 +22,12 @@ for custom `AtProtoRecord` classes.
 * Add "cashtag" facet detection to the default facet extractor. See [Add cashtag support for stock ticker discussions](https://github.com/bluesky-social/social-app/pull/9689)
 * Add `SetLiveStatus` to indicate if a user is live streaming, and `SetStatus` and `DeleteStatus` to set and delete a user's status.
 
+### Changed
+
+#### idunno.AtProto
+
+* Fix `PutRecord<TRecord>(AtProtoRepositoryRecord<TRecord>)` to use the record CID for `swapRecord`, not the `swapCommit`.
+
 ## 1.3.0 - 2026-01-26
 
 ### Added